substitute SsdTestShellApp
